requirements.txt
#requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-unstructured
langchain-huggingface
langchain-ollama
sentence-transformers
transformers
huggingface-hub
python-dotenv
pymupdf
python-docx
gitpython
python-multipart
unstructured
unstructured[pdf]
python-magic
pydantic
qdrant-client
langchain-qdrant
